requirements.txt
Pillow>=9.0.0
numpy>=1.24
opencv-python-headless>=4.8
pytesseract>=0.3.10
anyio>=3.7
sniffio>=1.3
scikit-image>=0.20
